@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -27,55 +27,42 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  carousel_slider: ^4.0.0
+  chips_choice_null_safety: ^2.0.6
+  cupertino_icons: ^1.0.2
+  dots_indicator: ^2.1.0
+  dropdown_button2: ^1.7.1
+  expandable: ^5.0.1
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  carousel_slider: ^4.0.0
-  dots_indicator: ^2.1.0
-  smooth_page_indicator: ^1.0.0+2
-  flutter_switch: ^0.3.2
-  qr_code_scanner: ^1.0.0
-  percent_indicator: ^4.2.2
-  flutter_svg: ^1.1.1+1
-  shared_preferences: ^2.0.15
-  http: ^0.13.4
-  toggle_switch: ^2.0.1
-  pin_code_fields: ^7.4.0
-  modal_bottom_sheet: ^2.1.0
   flutter_slidable: ^2.0.0
-  provider: ^6.0.3
-  expandable: ^5.0.1
+  flutter_svg: ^1.1.1+1
+  flutter_switch: ^0.3.2
+  geocoding: ^2.0.4
+  geolocator: ^9.0.1
+  http: ^0.13.4
   intl: ^0.17.0
+  modal_bottom_sheet: ^2.1.0
+  percent_indicator: ^4.2.2
+  pin_code_fields: ^7.4.0
+  provider: ^6.0.3
   pusher_channels_flutter: ^2.0.0
-  dropdown_button2: ^1.7.1
-  chips_choice_null_safety: ^2.0.6
-  
-  # pusher_channels_flutter: ^2.0.0
-  
-  
+  qr_code_scanner: ^1.0.0
+  qr_flutter: ^4.0.0
+  shared_preferences: ^2.0.15
+  smooth_page_indicator: ^1.0.0+2
+  toggle_switch: ^2.0.1
+  url_launcher: ^6.1.5
 
 dev_dependencies:
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
 
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^2.0.0
-  
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -85,14 +72,11 @@ flutter:
   assets:
     - assets/images/
     #- images/a_dot_burr.jpeg
-    
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
