@@ -1,3 +1,5 @@
+import 'image_model.dart';
+
 class GetAllCategoriesModel {
   bool? status;
   int? responseCode;
@@ -111,44 +113,122 @@ class SearchModel {
 class SearchData {
   int? id;
   String? name;
-  String? image;
-  String? parentId;
+  String? discount;
+  String? type;
+  String? price;
+  String? additionalDiscount;
+  String? additionalDiscountDate;
+  String? discountOnPrice;
+  String? afterDiscount;
+  String? actualPrice;
+  String? categoryId;
+  String? description;
   String? status;
+  String? rejectReason;
+  String? active;
+  String? activiationRequest;
+  String? merchantId;
+  String? createdBy;
   String? createdAt;
   String? updatedAt;
-  String? imagePath;
+  String? expiry;
+  String? merchantName;
+  String? categoryName;
+  ImageModel? image;
+  int? dealIsExpired;
+  String? typeName;
+  String? activationRequestFor;
 
   SearchData(
       {this.id,
       this.name,
-      this.image,
-      this.parentId,
+      this.discount,
+      this.type,
+      this.price,
+      this.additionalDiscount,
+      this.additionalDiscountDate,
+      this.discountOnPrice,
+      this.afterDiscount,
+      this.actualPrice,
+      this.categoryId,
+      this.description,
       this.status,
+      this.rejectReason,
+      this.active,
+      this.activiationRequest,
+      this.merchantId,
+      this.createdBy,
       this.createdAt,
       this.updatedAt,
-      this.imagePath});
+      this.expiry,
+      this.merchantName,
+      this.categoryName,
+      this.image,
+      this.dealIsExpired,
+      this.typeName,
+      this.activationRequestFor});
 
   SearchData.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     name = json['name'];
-    image = json['image'];
-    parentId = json['parent_id'];
+    discount = json['discount'];
+    type = json['type'];
+    price = json['price'];
+    additionalDiscount = json['additional_discount'];
+    additionalDiscountDate = json['additional_discount_date'];
+    discountOnPrice = json['discount_on_price'];
+    afterDiscount = json['after_discount'];
+    actualPrice = json['actual_price'];
+    categoryId = json['category_id'];
+    description = json['description'];
     status = json['status'];
+    rejectReason = json['reject_reason'];
+    active = json['active'];
+    activiationRequest = json['activiation_request'];
+    merchantId = json['merchant_id'];
+    createdBy = json['created_by'];
     createdAt = json['created_at'];
     updatedAt = json['updated_at'];
-    imagePath = json['imagePath'];
+    expiry = json['expiry'];
+    merchantName = json['merchant_name'];
+    categoryName = json['category_name'];
+    image = json['image'] != null ? ImageModel.fromJson(json['image']) : null;
+    dealIsExpired = json['dealIsExpired'];
+    typeName = json['TypeName'];
+    activationRequestFor = json['activationRequestFor'];
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
+    final Map<String, dynamic> data = Map<String, dynamic>();
     data['id'] = id;
     data['name'] = name;
-    data['image'] = image;
-    data['parent_id'] = parentId;
+    data['discount'] = discount;
+    data['type'] = type;
+    data['price'] = price;
+    data['additional_discount'] = additionalDiscount;
+    data['additional_discount_date'] = additionalDiscountDate;
+    data['discount_on_price'] = discountOnPrice;
+    data['after_discount'] = afterDiscount;
+    data['actual_price'] = actualPrice;
+    data['category_id'] = categoryId;
+    data['description'] = description;
     data['status'] = status;
+    data['reject_reason'] = rejectReason;
+    data['active'] = active;
+    data['activiation_request'] = activiationRequest;
+    data['merchant_id'] = merchantId;
+    data['created_by'] = createdBy;
     data['created_at'] = createdAt;
     data['updated_at'] = updatedAt;
-    data['imagePath'] = imagePath;
+    data['expiry'] = expiry;
+    data['merchant_name'] = merchantName;
+    data['category_name'] = categoryName;
+    if (image != null) {
+      data['image'] = image!.toJson();
+    }
+    data['dealIsExpired'] = dealIsExpired;
+    data['TypeName'] = typeName;
+    data['activationRequestFor'] = activationRequestFor;
     return data;
   }
 }
