@@ -1,26 +1,104 @@
+import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter/material.dart';
 import 'package:gigi_app/screens/full_menu/profile.dart';
+import 'package:gigi_app/services/branch/branch_services.dart';
 
-class location_bar extends StatelessWidget {
-  const location_bar({Key? key, required this.token, this.city})
-      : super(key: key);
+class location_bar extends StatefulWidget {
+  const location_bar({Key? key, required this.token}) : super(key: key);
   final String token;
-  final String? city;
+
+  @override
+  State<location_bar> createState() => _location_barState();
+}
+
+class _location_barState extends State<location_bar> {
+  List<dynamic>? items = List.empty(growable: true);
+  List<dynamic>? data;
+  String? selectedValue;
+  Future<List> getAllBrachesDropdown() async {
+    final result = await BranchServices().getBranchesList(token: widget.token);
+    for (var i in result['data']) {
+      items!.add(i['name']);
+      print(items);
+    }
+    return items!;
+  }
+
+  @override
+  void initState() {
+    getAllBrachesDropdown().whenComplete(() {
+      print('data : $data');
+    }).then((value) {
+      setState(() {
+        data = value;
+      });
+    });
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
     return Row(
       children: [
         Image.asset('assets/images/Vector.png'),
-        const Padding(
-          padding: EdgeInsets.only(left: 10.94),
-          child: Text('Branch 1',
-              style: TextStyle(
-                fontFamily: 'Mulish',
-                fontSize: 16,
-                fontWeight: FontWeight.w700,
-                color: Color(0xFF0D9BFF),
-              )),
+        Padding(
+          padding: const EdgeInsets.only(left: 10.94),
+          child: data == null
+              ? const Text('fetching...')
+              : DropdownButton2(
+                  icon: const SizedBox(),
+                  isExpanded: true,
+                  hint: Text(
+                    data!.isEmpty ? 'no branch' : data![0]!,
+                    style: const TextStyle(
+                        fontFamily: 'Mulish',
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                        color: Color(0xFF0D9BFF)),
+                    overflow: TextOverflow.ellipsis,
+                  ),
+                  items: data!
+                      .map((item) => DropdownMenuItem<String>(
+                            value: item,
+                            child: Text(
+                              item,
+                              style: const TextStyle(
+                                  fontFamily: 'Mulish',
+                                  fontSize: 16,
+                                  fontWeight: FontWeight.w700,
+                                  color: Color(0xFF0D9BFF)),
+                              overflow: TextOverflow.ellipsis,
+                            ),
+                          ))
+                      .toList(),
+                  value: selectedValue,
+                  onChanged: (value) {
+                    setState(() {
+                      selectedValue = value as String;
+                    });
+                  },
+                  underline: const SizedBox(),
+                  buttonHeight: 30,
+                  buttonWidth: 200,
+                  buttonPadding: const EdgeInsets.only(left: 14, right: 8),
+                  buttonDecoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(14),
+                  ),
+                  buttonElevation: 0,
+                  itemHeight: 40,
+                  itemPadding: const EdgeInsets.only(left: 14, right: 14),
+                  dropdownMaxHeight: 200,
+                  dropdownWidth: 160,
+                  dropdownPadding: null,
+                  dropdownDecoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(14),
+                  ),
+                  dropdownElevation: 8,
+                  scrollbarRadius: const Radius.circular(40),
+                  scrollbarThickness: 6,
+                  scrollbarAlwaysShow: true,
+                  offset: const Offset(-20, 0),
+                ),
         ),
         const Spacer(),
         InkWell(
@@ -28,7 +106,7 @@ class location_bar extends StatelessWidget {
           onTap: () {
             Navigator.of(context).push(MaterialPageRoute(
                 builder: (context) => Profile(
-                      token: token,
+                      token: widget.token,
                       isLeadingIcon: true,
                     )));
           },
