@@ -1,13 +1,33 @@
+import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter/material.dart';
+import 'package:gigi_app/services/deals/user_deals_services.dart';
 import 'package:gigi_app/user_app/notification_screen.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../user_app/user_menu/ham_user.dart';
 
-class Location_bar_user extends StatelessWidget {
-  const Location_bar_user({Key? key, required this.token, this.city})
+class Location_bar_user extends StatefulWidget {
+  const Location_bar_user(
+      {Key? key, required this.token, this.updateTrendingList})
       : super(key: key);
   final String token;
-  final String? city;
+  final FutureBuilder? updateTrendingList;
+
+  @override
+  State<Location_bar_user> createState() => _Location_bar_userState();
+}
+
+class _Location_bar_userState extends State<Location_bar_user> {
+  String? selectedValue;
+  String? country;
+
+  @override
+  void initState() {
+    super.initState();
+    // getCountry().whenComplete(() {
+    fetchCitiesAndCountries();
+    // });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -16,20 +36,69 @@ class Location_bar_user extends StatelessWidget {
         Image.asset('assets/images/Vector.png'),
         Padding(
           padding: const EdgeInsets.only(left: 10.94),
-          child: Text(city ?? 'Baku, Azerbaijan',
-              style: const TextStyle(
-                fontFamily: 'Mulish',
-                fontSize: 16,
-                fontWeight: FontWeight.w700,
-                color: Color(0xFF0D9BFF),
-              )),
+          child: items == null
+              ? const Text('fetching...')
+              : DropdownButton2(
+                  icon: const SizedBox(),
+                  isExpanded: true,
+                  hint: Text(
+                    country == null || country!.isEmpty ? 'Country' : country!,
+                    style: const TextStyle(
+                        fontFamily: 'Mulish',
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                        color: Color(0xFF0D9BFF)),
+                    overflow: TextOverflow.ellipsis,
+                  ),
+                  items: items!
+                      .map((item) => DropdownMenuItem<String>(
+                            value: item,
+                            child: Text(
+                              item,
+                              style: const TextStyle(
+                                  fontFamily: 'Mulish',
+                                  fontSize: 16,
+                                  fontWeight: FontWeight.w700,
+                                  color: Color(0xFF0D9BFF)),
+                              overflow: TextOverflow.ellipsis,
+                            ),
+                          ))
+                      .toList(),
+                  value: selectedValue,
+                  onChanged: (value) {
+                    setState(() {
+                      selectedValue = value as String;
+                    });
+                  },
+                  underline: const SizedBox(),
+                  buttonHeight: 30,
+                  buttonWidth: 100,
+                  buttonPadding: const EdgeInsets.only(left: 14, right: 8),
+                  buttonDecoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(14),
+                  ),
+                  buttonElevation: 0,
+                  itemHeight: 40,
+                  itemPadding: const EdgeInsets.only(left: 14, right: 14),
+                  dropdownMaxHeight: 200,
+                  dropdownWidth: 160,
+                  dropdownPadding: null,
+                  dropdownDecoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(14),
+                  ),
+                  dropdownElevation: 8,
+                  scrollbarRadius: const Radius.circular(40),
+                  scrollbarThickness: 6,
+                  scrollbarAlwaysShow: true,
+                  offset: const Offset(-20, 0),
+                ),
         ),
         const Spacer(),
         GestureDetector(
             onTap: () {
               Navigator.of(context).push(MaterialPageRoute(
                   builder: (_) => ham_user(
-                        token: token,
+                        token: widget.token,
                       )));
             },
             child: Image.asset('assets/images/drawer.png')),
@@ -37,7 +106,8 @@ class Location_bar_user extends StatelessWidget {
         InkWell(
           onTap: () {
             Navigator.of(context).push(MaterialPageRoute(
-                builder: ((context) => NotificationScreen(token: token))));
+                builder: ((context) =>
+                    NotificationScreen(token: widget.token))));
           },
           child: Image.asset(
             'assets/images/notification.png',
@@ -47,5 +117,21 @@ class Location_bar_user extends StatelessWidget {
         )
       ],
     );
+  }
+
+  Future<void> getCountry() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    country = prefs.getString('country');
+  }
+
+  List<dynamic>? items;
+
+  Future<void> fetchCitiesAndCountries() async {
+    final result =
+        await UserDealServices().getSystemCities(widget.token, 'Pakistan');
+
+    setState(() {
+      items = result['data'];
+    });
   }
 }
