@@ -3,8 +3,8 @@ import 'package:gigi_app/providers/deal_provider.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:provider/provider.dart';
 
+import '../../apis/api_urls.dart';
 import '../../constant/size_constants.dart';
-import '../../models/reviews_model.dart';
 import 'deals_details.dart';
 
 class dealsUser extends StatefulWidget {
@@ -12,8 +12,6 @@ class dealsUser extends StatefulWidget {
       : super(key: key);
   final String dealData;
   final String token;
-
-  static const url = 'https://gigiapi.zanforthstaging.com/';
 
   @override
   State<dealsUser> createState() => _dealsUserState();
@@ -145,18 +143,20 @@ class _dealsUserState extends State<dealsUser> {
                                 children: [
                                   Image.asset('assets/images/rating.png',
                                       width: 6, height: 6),
-                                  Text(
-                                    Reviews().getRating(
-                                        dealProvider!.dealData.reviews),
-                                    style: const TextStyle(
+                                  const Text(
+                                    '0',
+                                    // Reviews().getRating(
+                                    //     dealProvider!.dealData.reviews),
+                                    style: TextStyle(
                                         fontFamily: 'Poppins',
                                         fontSize: 7,
                                         fontWeight: FontWeight.w700,
                                         color: Color(0xFFFFFFFF)),
                                   ),
-                                  Text(
-                                    '(${dealProvider!.dealData.reviews!.length} reviews)',
-                                    style: const TextStyle(
+                                  const Text(
+                                    '0 reviews',
+                                    // '(${dealProvider!.dealData.reviews!.length} reviews)',
+                                    style: TextStyle(
                                         fontFamily: 'Poppins',
                                         fontSize: 4,
                                         fontWeight: FontWeight.w400,
@@ -229,11 +229,11 @@ class _dealsUserState extends State<dealsUser> {
                       ),
                       const Spacer(),
                       Expanded(
-                        child: dealProvider!.dealData.images == null ||
-                                dealProvider!.dealData.images!.isEmpty
+                        child: dealProvider!.dealData.image == null ||
+                                dealProvider!.dealData.image!.image!.isEmpty
                             ? Image.asset('assets/images/food.png')
                             : Image.network(
-                                '${dealsUser.url}${dealProvider!.dealData.images![0].path}/${dealProvider!.dealData.images![0].image}',
+                                '${ApiUrls.imgBaseUrl}${dealProvider!.dealData.image!.path}/${dealProvider!.dealData.image!.image}',
                               ),
                       ),
                     ],
