@@ -3,6 +3,8 @@ import 'package:gigi_app/models/merchant_model.dart';
 import 'package:gigi_app/services/user_merchant_services.dart';
 import 'package:gigi_app/user_app/user_menu/user_review.dart';
 
+import '../../apis/api_urls.dart';
+
 class StoreProfile extends StatelessWidget {
   const StoreProfile({Key? key, required this.profileId, required this.token})
       : super(key: key);
@@ -32,172 +34,192 @@ class StoreProfile extends StatelessWidget {
               child: SingleChildScrollView(
                 child: FutureBuilder<SingleMerchant>(
                   future: UserMerchantServices()
-                      .singleMerchantProfile(id: '3', token: token),
+                      .singleMerchantProfile(id: profileId, token: token),
                   builder: ((context, snapshot) {
                     List<Widget> children;
                     if (snapshot.hasData) {
                       var data = snapshot.data!.data!;
-                      children = [
-                        Image.asset('assets/images/slt.png'),
-                        Padding(
-                          padding: const EdgeInsets.only(top: 13, bottom: 10),
-                          child: Text(data.name!,
-                              style: const TextStyle(
-                                  fontFamily: 'DMSans',
-                                  fontSize: 22,
-                                  fontWeight: FontWeight.w500,
-                                  color: Color(0xff000000))),
-                        ),
-                        Text(
-                          data.branches![0].address!,
-                          style: const TextStyle(
+                      if (data.statusName != 'Active') {
+                        children = [
+                          Center(
+                            child: Text(
+                                '${snapshot.data!.data!.name} is not active'),
+                          ),
+                        ];
+                      } else {
+                        children = [
+                          data.profilePicture == null
+                              ? Image.asset('assets/images/slt.png')
+                              : Image.network(
+                                  '${ApiUrls.imgBaseUrl}/${data.profilePicturePath}/${data.profilePicture}'),
+                          Padding(
+                            padding: const EdgeInsets.only(top: 13, bottom: 10),
+                            child: Text(data.name!,
+                                style: const TextStyle(
+                                    fontFamily: 'DMSans',
+                                    fontSize: 22,
+                                    fontWeight: FontWeight.w500,
+                                    color: Color(0xff000000))),
+                          ),
+                          Text(
+                            data.branches == null || data.branches!.isEmpty
+                                ? 'merchant have no branch'
+                                : data.branches![0].address ??
+                                    'no address found',
+                            style: const TextStyle(
+                                fontFamily: 'DMSans',
+                                fontSize: 10,
+                                fontWeight: FontWeight.w500,
+                                color: Color(0xff172995)),
+                            textAlign: TextAlign.center,
+                          ),
+                          const SizedBox(
+                            height: 5,
+                          ),
+                          Text(
+                            '${data.phone ?? 'not povided'} - ${data.email}',
+                            style: const TextStyle(
                               fontFamily: 'DMSans',
                               fontSize: 10,
                               fontWeight: FontWeight.w500,
-                              color: Color(0xff172995)),
-                          textAlign: TextAlign.center,
-                        ),
-                        const SizedBox(
-                          height: 5,
-                        ),
-                        Text(
-                          '${data.phone} - ${data.email}',
-                          style: const TextStyle(
-                            fontFamily: 'DMSans',
-                            fontSize: 10,
-                            fontWeight: FontWeight.w500,
-                            color: Color(0xff000000),
+                              color: Color(0xff000000),
+                            ),
+                            textAlign: TextAlign.center,
                           ),
-                          textAlign: TextAlign.center,
-                        ),
-                        const SizedBox(
-                          height: 5,
-                        ),
-                        Row(
+                          const SizedBox(
+                            height: 5,
+                          ),
+                          Row(
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              children: [
+                                Text(
+                                  '${data.averageRating ?? 0}',
+                                  style: const TextStyle(
+                                      fontSize: 30, color: Color(0xff172995)),
+                                  textAlign: TextAlign.center,
+                                ),
+                                const SizedBox(width: 5),
+                                Column(
+                                  children: [
+                                    Row(
+                                      children: [
+                                        Image.asset('assets/images/rate.png',
+                                            width: 18, height: 18),
+                                        Image.asset('assets/images/rate.png',
+                                            width: 18, height: 18),
+                                        Image.asset('assets/images/rate.png',
+                                            width: 18, height: 18),
+                                        Image.asset('assets/images/rate.png',
+                                            width: 18, height: 18),
+                                        Image.asset('assets/images/rate.png',
+                                            width: 18, height: 18),
+                                      ],
+                                    ),
+                                    const Text(
+                                      'Reviews and Rating',
+                                      style: TextStyle(
+                                        fontFamily: 'DMSans',
+                                        fontSize: 10,
+                                        fontWeight: FontWeight.w500,
+                                        color: Color(0xff000000),
+                                      ),
+                                      textAlign: TextAlign.center,
+                                    ),
+                                  ],
+                                )
+                              ]),
+                          const SizedBox(height: 10),
+                          Container(
+                            height: 30,
+                            width: 70,
+                            decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(10),
+                                color: const Color(0xff172995)),
+                            child: const Center(
+                              child: Text('View',
+                                  style: TextStyle(
+                                      fontFamily: 'DMSans',
+                                      fontSize: 20,
+                                      fontWeight: FontWeight.w500,
+                                      color: Color(0xffFCFCFC))),
+                            ),
+                          ),
+                          const SizedBox(height: 20),
+                          Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                const Text('48',
+                                    style: TextStyle(
+                                        fontFamily: 'DMSans',
+                                        fontSize: 30,
+                                        fontWeight: FontWeight.w500,
+                                        color: Color(0xff172995))),
+                                Image.asset('assets/images/divide.png',
+                                    width: 30, height: 30),
+                                const Text('400+',
+                                    style: TextStyle(
+                                        fontFamily: 'DMSans',
+                                        fontSize: 30,
+                                        fontWeight: FontWeight.w500,
+                                        color: Color(0xff172995))),
+                                Image.asset('assets/images/divide.png',
+                                    width: 30, height: 30),
+                                const Text('2',
+                                    style: TextStyle(
+                                        fontFamily: 'DMSans',
+                                        fontSize: 30,
+                                        fontWeight: FontWeight.w500,
+                                        color: Color(0xff172995))),
+                              ]),
+                          const SizedBox(height: 20),
+                          const Text('Reviews',
+                              style: TextStyle(
+                                  fontFamily: 'DMSans',
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.w500,
+                                  color: Color(0xff000000))),
+                          const SizedBox(height: 10),
+                          Row(
                             mainAxisAlignment: MainAxisAlignment.center,
                             children: [
-                              Text(
-                                '${data.averageRating ?? 0}',
-                                style: const TextStyle(
-                                    fontSize: 30, color: Color(0xff172995)),
-                                textAlign: TextAlign.center,
+                              Image.asset('assets/images/rate.png',
+                                  width: 18, height: 18),
+                              Image.asset('assets/images/rate.png',
+                                  width: 18, height: 18),
+                              Image.asset('assets/images/rate.png',
+                                  width: 18, height: 18),
+                              Image.asset('assets/images/rate.png',
+                                  width: 18, height: 18),
+                              Image.asset('assets/images/rate.png',
+                                  width: 18, height: 18),
+                              const SizedBox(
+                                width: 20,
                               ),
-                              const SizedBox(width: 5),
-                              Column(
-                                children: [
-                                  Row(
-                                    children: [
-                                      Image.asset('assets/images/rate.png',
-                                          width: 18, height: 18),
-                                      Image.asset('assets/images/rate.png',
-                                          width: 18, height: 18),
-                                      Image.asset('assets/images/rate.png',
-                                          width: 18, height: 18),
-                                      Image.asset('assets/images/rate.png',
-                                          width: 18, height: 18),
-                                      Image.asset('assets/images/rate.png',
-                                          width: 18, height: 18),
-                                    ],
-                                  ),
-                                  const Text(
-                                    'Reviews and Rating',
-                                    style: TextStyle(
+                              Text('${data.reviews!.length} Reviews',
+                                  style: const TextStyle(
                                       fontFamily: 'DMSans',
                                       fontSize: 10,
                                       fontWeight: FontWeight.w500,
-                                      color: Color(0xff000000),
-                                    ),
-                                    textAlign: TextAlign.center,
-                                  ),
-                                ],
-                              )
-                            ]),
-                        const SizedBox(height: 10),
-                        Container(
-                          height: 30,
-                          width: 70,
-                          decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(10),
-                              color: const Color(0xff172995)),
-                          child: const Center(
-                            child: Text('View',
-                                style: TextStyle(
-                                    fontFamily: 'DMSans',
-                                    fontSize: 20,
-                                    fontWeight: FontWeight.w500,
-                                    color: Color(0xffFCFCFC))),
+                                      color: Color(0xff898989)))
+                            ],
                           ),
-                        ),
-                        const SizedBox(height: 20),
-                        Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              const Text('48',
-                                  style: TextStyle(
-                                      fontFamily: 'DMSans',
-                                      fontSize: 30,
-                                      fontWeight: FontWeight.w500,
-                                      color: Color(0xff172995))),
-                              Image.asset('assets/images/divide.png',
-                                  width: 30, height: 30),
-                              const Text('400+',
-                                  style: TextStyle(
-                                      fontFamily: 'DMSans',
-                                      fontSize: 30,
-                                      fontWeight: FontWeight.w500,
-                                      color: Color(0xff172995))),
-                              Image.asset('assets/images/divide.png',
-                                  width: 30, height: 30),
-                              const Text('2',
-                                  style: TextStyle(
-                                      fontFamily: 'DMSans',
-                                      fontSize: 30,
-                                      fontWeight: FontWeight.w500,
-                                      color: Color(0xff172995))),
-                            ]),
-                        const SizedBox(height: 20),
-                        const Text('Reviews',
-                            style: TextStyle(
-                                fontFamily: 'DMSans',
-                                fontSize: 20,
-                                fontWeight: FontWeight.w500,
-                                color: Color(0xff000000))),
-                        const SizedBox(height: 10),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Image.asset('assets/images/rate.png',
-                                width: 18, height: 18),
-                            Image.asset('assets/images/rate.png',
-                                width: 18, height: 18),
-                            Image.asset('assets/images/rate.png',
-                                width: 18, height: 18),
-                            Image.asset('assets/images/rate.png',
-                                width: 18, height: 18),
-                            Image.asset('assets/images/rate.png',
-                                width: 18, height: 18),
-                            const SizedBox(
-                              width: 20,
-                            ),
-                            Text('${data.reviews!.length} Reviews',
-                                style: const TextStyle(
-                                    fontFamily: 'DMSans',
-                                    fontSize: 10,
-                                    fontWeight: FontWeight.w500,
-                                    color: Color(0xff898989)))
-                          ],
-                        ),
-                        if (data.reviews != null || data.reviews!.isNotEmpty)
-                          ListView.builder(
-                              itemCount: data.reviews!.length,
-                              shrinkWrap: true,
-                              itemBuilder: (context, index) {
-                                return user_review(
-                                  review: data.reviews![index].notes!,
-                                  reviewSender: data.reviews![index].userName!,
-                                );
-                              })
-                      ];
+                          data.reviews == null || data.reviews!.isEmpty
+                              ? const Center(child: Text('Not reviewed yet'))
+                              : ListView.builder(
+                                  itemCount: data.reviews!.length,
+                                  shrinkWrap: true,
+                                  itemBuilder: (context, index) {
+                                    return user_review(
+                                      review: data.reviews![index].notes == null
+                                          ? ''
+                                          : data.reviews![index].notes ?? '',
+                                      reviewSender:
+                                          data.reviews![index].userName ??
+                                              'no name',
+                                    );
+                                  })
+                        ];
+                      }
                     } else if (snapshot.hasError) {
                       children = <Widget>[
                         const Icon(
